map: Add get with precomputed hash function
A user of the cerve_map API may have an item they look up very often.
This function allows them to compute the hash for the key once and reuse
it during lookups. It is up to the user of the function to ensure the
key actually maps to that hash.

NOTE: If the precomputed hash is rarely used, this will likely result in
cache misses and end up taking longer than the hashing. This should be
used carefully.