map: Add other precomp_hash fn variants & don't require free_data
precomp_hash function allow the user to compute the hash once and reuse
it in mulitple calls. Hashing isn't very expensive but might as well
save all the time you can.

The free_data callback isn't always necessary. By not requiring it,
the deinit path can be much faster (O(1) instead of O(n)) if it is
not needed.