http headers: Implement structures for request HTTP headers
There are two major cases when dealing with HTTP headers:
1. Read heavy (parsing requests)
2. Write heavy (creating responses)
Cerve will implement two HTTP header struct variants.

The read heavy structure will be a flat buffer that is "indexed" with a
hashmap. This allows a user to quickly get the values for an HTTP
header. Building the map is expensive.

The write heavy structure will be a flat buffer that can be directly
sent over TCP. A user will almost never need to read an HTTP header
when creating responses. And if that does become a real use case, we
can merge the read and write heavy implementations and keep a map
that points into the raw buffer.